video_subtitles/requirements.txt
langchain==0.1.0
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
moviepy>=1.0.3
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0